# feat(ext/web): resourceForReadableStream