… Cargo.lock
… cli/tests/unit/streams_test.ts
… ext/web/stream_resource.rs